% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_match_score.R
\name{get_match_score}
\alias{get_match_score}
\title{Return match scores by attribute}
\usage{
get_match_score(r_mat, s_mat, attr)
}
\arguments{
\item{r_mat}{A matrix of research project needs}

\item{s_mat}{A matrix of student interest}

\item{attr}{A character vector with the attribute describing which match to evluate (datatype, skill, course, discipline)}
}
\value{
A rectangular matrix of project x student with correlation values for different attributes
}
\description{
Return match scores by attribute
}
\examples{
course_match <- get_match_score(attr = 'course')

}
