% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_index.R
\name{get_project_index}
\alias{get_project_index}
\title{Pull researcher index}
\usage{
get_project_index(project_url)
}
\arguments{
\item{project_url}{A URL to the Google Sheet where research project survey responses are stored}
}
\value{
A data frame of all submitted research projects
}
\description{
Pull researcher index
}
\examples{
p_mat <- get_project_index(project_url)

}
