% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_student_index.R
\name{get_student_index}
\alias{get_student_index}
\title{Pull student index}
\usage{
get_student_index(student_url)
}
\arguments{
\item{project_url}{A URL to the Google Sheet where student interest survey responses are stored}
}
\value{
A data frame of all submitted student applications
}
\description{
Pull student index
}
\examples{
s_mat <- get_student_index(student_url)

}
