% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_top_matches.R
\name{identify_top_matches}
\alias{identify_top_matches}
\title{Identify top students for project}
\usage{
identify_top_matches(r_mat, s_mat)
}
\arguments{
\item{r_mat}{A matrix of research project needs}

\item{s_mat}{A matrix of student interest}
}
\value{
List of projects, each of which contains a data frame with top matched students by four criteria and their scores, where scores closer to zero are better
}
\description{
Identify top students for project
}
\examples{
proj_list <- identify_top_matches(r_mat, s_mat)

}
