% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_project_needs.R
\name{pull_research_needs}
\alias{pull_research_needs}
\title{Pull research project description}
\usage{
pull_research_needs(project_url, student_url)
}
\value{
A rectangular matrix of researcher responses and scores
}
\description{
Pull research project description
}
\examples{
p_mat <- pull_project_needs()

}
