% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_student_interest.R
\name{pull_student_interest}
\alias{pull_student_interest}
\title{Pull student interest forms}
\usage{
pull_student_interest(student_url)
}
\value{
A rectangular matrix of student responses and scores
}
\description{
Pull student interest forms
}
\examples{
s_mat <- pull_student_interest()

}
