% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_to_df.R
\name{transcript_to_df}
\alias{transcript_to_df}
\title{Extract course-related data from PDF transcript}
\usage{
transcript_to_df(file)
}
\arguments{
\item{file}{The filepath to the PDF transcript}
}
\value{
A data frame of course codes, names, and grades
}
\description{
Extract course-related data from PDF transcript
}
\examples{
df <- transcript_to_df(file = "transcript.pdf")

}
