% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UCDdept}
\alias{UCDdept}
\title{Course codes/departments for UCD}
\format{
A data frame with 220 rows and 2 variables
}
\source{
\url{https://registrar.ucdavis.edu/records/transcripts/subject-codes}
}
\usage{
UCDdept
}
\description{
A dats frame with course codes and their associated name
}
\keyword{datasets}
