% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_top_matches.R
\name{identify_top_matches}
\alias{identify_top_matches}
\title{Identify top students for project}
\usage{
identify_top_matches(p_mat, s_mat, cutoff)
}
\arguments{
\item{p_mat}{A matrix of research project needs}

\item{s_mat}{A matrix of student interest}

\item{cutoff}{A value 0-3 identifying a lower bound threshold for selecting students. For example, selecting a cutoff value of 2 will show students with match scores >= 2 for each project. cutoff = 0 will display all students, cutoff = 3 will display only perfect matches.}
}
\value{
List of projects, each of which contains a data frame with top matched students by four criteria and their scores, where scores closer to zero are better
}
\description{
Identify top students for project
}
\examples{
proj_list <- identify_top_matches(p_mat, s_mat)

}
